pyproject.toml
[project]
name = "home-credit-default-risk-ml"
version = "0.1.0"
description = "Leakage-safe Home Credit default-risk machine-learning pipeline"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "catboost>=1.2.10",
    "dagshub>=0.7.0",
    "dotenv>=0.9.9",
    "dvc>=3.67.1",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "imbalanced-learn>=0.14.1",
    "ipykernel>=7.2.0",
    "joblib>=1.5.3",
    "kaggle>=2.1.0",
    "kagglehub>=1.0.0",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "missingno>=0.5.2",
    "mlflow>=3.10.1,<3.11",
    "numpy>=2.4.4",
    "optuna>=4.8.0",
    "pandas>=2.2,<3",
    "polars>=1.40.0",
    "pydantic>=2.12.0",
    "pyarrow>=20,<23",
    "pytest>=9.0.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "tqdm>=4.67.3",
    "uvicorn>=0.38.0",
    "xgboost>=3.2.0",
]
